// src/types/usuario.ts
export interface UsuarioResponse {
  id: number;
  username: string;
  nome: string;
  email: string;
  cpf: string;
  dataNascimento: string;
  telefone: string;
  cep: string;
  endereco: string;
  complemento: string;
  cidade: string;
  uf: string;
  createdAt: string;
  updatedAt: string;
  isAdmin: boolean;
}